# api_rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto DAW Backend API",
        "description": "API REST del Proyecto DAW. Esta API gestiona la lógica de negocio para la plataforma de intercambio de tiempo.\n *     \n *     Módulos incluidos:\n *     - **Autenticación**: Registro, login, logout y gestión de perfil.\n *     - **Usuarios**: Gestión completa de usuarios (CRUD).\n *     - **Servicios**: Gestión de ofertas y demandas de servicios.\n *     - **Transacciones**: Control de intercambios de tiempo entre usuarios.\n *     - **Valoraciones**: Sistema de reputación y feedback.\n *     - **Mensajes**: Sistema de mensajería interna.\n *     - **Common**: Recursos compartidos (provincias, poblaciones, etc.).",
        "contact": {
            "email": "soporte@proyectodaw.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://proyecto-daw-backend.onrender.com/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener todos los servicios",
                "operationId": "50ae03f33a22e845432a3317903b291f",
                "responses": {
                    "200": {
                        "description": "Lista de servicios obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Todos los servicios obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 19
                                                    },
                                                    "usuario_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "categoria_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "tipo": {
                                                        "type": "string",
                                                        "example": "oferta"
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Clases de matemáticas"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Ofrezco clases particulares de matemáticas"
                                                    },
                                                    "provincia": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "ciudad": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "horas_estimadas": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "activo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "0bf35b8fa344e5b2563343ba426001f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "categoria_id",
                                    "tipo",
                                    "titulo",
                                    "descripcion",
                                    "provincia_id",
                                    "ciudad_id",
                                    "horas_estimadas"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "description": "oferta o demanda",
                                        "type": "string",
                                        "example": "oferta"
                                    },
                                    "titulo": {
                                        "type": "string",
                                        "example": "Clases de matemáticas"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Ofrezco clases particulares de matemáticas para ESO y Bachillerato"
                                    },
                                    "provincia_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "ciudad_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "horas_estimadas": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "estado": {
                                        "description": "activo, en_proceso, finalizado o cancelado",
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/servicios/{user_id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener todos los servicios de un usuario",
                "operationId": "4014be8e1d1fea7dc1407ecfc402aa31",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicios del usuario obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servicios del usuario obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 406
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación mediante token Laravel Sanctum. Ingrese el token directamente.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}